test/alternator: move the shared S3 helpers into util.py
test_export.py grew a small set of S3 helpers - unique_bucket_name(),
make_s3_client(), block_bucket_writes_on_s3() and the new_s3_bucket()
context manager. The import-from-S3 tests need exactly the same plumbing,
so move it to util.py.
